Vision.Image.Storage.DevIL

Description

Uses the DevIL C library to read and write images from and to files and memory buffers.

Please read our README to to get a detailed usage and some examples.

Note: As the underlying DevIL library is *not* tread-safe, there is a global lock which will prevent two load/save calls to be performed at the same time. The Haskell interface should be thread-safe but will not be able to benefit from multiple processing cores.

Types & classes

data StorageImage

Constructors

GreyStorage Grey

RGBAStorage RGBA

RGBStorage RGB

data StorageError

Constructors

FailedToInit
Failed to initialise the library.

FailedToOpenFile
Failed to open the given file.

FailedToLoad
Failed to load the image, invalid format.

FailedToHaskell
Failed to convert the loaded image to its Haskell representation.

FailedToDevil
Failed to write the image content through the inner DevIL library.

FailedToSave
Failed to save the image.

FileAlreadyExists
The file already exists.

InvalidFile
The file could not be loaded because of an invalid value.

OutOfMemory
Could not allocate memory for the new image data.

UnknownFileType
The file content or extension does not match any known image type.

UnknownError (Maybe String)

class ImageType t

Minimal complete definition

toIlType

class ImageType t => LoadImageType t

Image types which can be loaded using load and loadBS.

class ImageType t => SaveImageType t

Image types which can be loaded using save.

class ImageType t => SaveBSImageType t

Image types which can be loaded using saveBS.

Functions

load :: (LoadImageType t, Convertible StorageImage i) => t -> FilePath -> IO (Either StorageError i)

Reads an image from a file.

If the Autodetect image type is given, type will be determined automatically with the file extension and the file headers.

loadBS :: (LoadImageType t, Convertible StorageImage i) => t -> ByteString -> Either StorageError i

Reads an image from a strict ByteString.

If the Autodetect image type is given, type will be determined automatically with the file headers.

save :: (SaveImageType t, Convertible i StorageImage) => t -> FilePath -> i -> IO (Maybe StorageError)

Saves the image to the given file.

If the Autodetect image type is given, type will be determined automatically with the file extension.

Note: will fail if the file already exists.

saveBS :: (SaveBSImageType t, Convertible i StorageImage) => t -> i -> Either StorageError ByteString

Saves the image into a manifest vector from a strict ByteString.

Images types

data Autodetect

Tries to determine automatically the image type with the header and/or the file extension.

Raises an UnknownFileType error if the detection failed.

Constructors

Autodetect

data BLP

Constructors

BLP

data BMP

Constructors

BMP

data CUT

Constructors

CUT

data DCX

Constructors

DCX

data DDS

DirectDraw Surface (.dds).

Constructors

DDS

data DICOM

Constructors

DICOM

data Doom

Doom texture.

Constructors

Doom

data DoomFlat

Doom flat texture (floor).

Constructors

DoomFlat

data DPX

Constructors

DPX

data EXR

Constructors

EXR

data FITS

Constructors

FITS

data FTX

Constructors

FTX

data GIF

Constructors

GIF

data HDR

Constructors

HDR

data ICO

Constructors

ICO

data ICNS

Constructors

ICNS

data IFF

Constructors

IFF

data IWI

Constructors

IWI

data JASCPAL

Paint Shop Pro (Jasc) palette.

Constructors

JASCPAL

data JP2

JPEG 2000.

Constructors

JP2

data JPG

Constructors

JPG

data LIF

Homeworld (.lif).

Constructors

LIF

data MDL

Constructors

MDL

data MNG

Constructors

MNG

data MP3

Constructors

MP3

data PCD

Constructors

PCD

data PCX

Constructors

PCX

data PIC

Constructors

PIC

data PIX

Constructors

PIX

data PNG

Constructors

PNG

data PNM

Portable AnyMap (.pbm, .pgm or .ppm).

Constructors

PNM

data PSD

Photoshop Document.

Constructors

PSD

data PSP

Paint Shop Pro image.

Constructors

PSP

data PXR

Constructors

PXR

data RAW

Raw data with a 13-byte header.

Constructors

RAW

data ROT

Constructors

ROT

data SGI

Constructors

SGI

data SUN

Constructors

SUN

data Texture

Medieval II: Total War Texture (.texture) file.

Constructors

Texture

data TGA

Constructors

TGA

data TIFF

Constructors

TIFF

data TPL

Constructors

TPL

data UTX

Constructors

UTX

data VTF

Constructors

VTF

data WAL

Constructors

WAL

data WBMP

Constructors

WBMP

data XPM

Constructors

XPM