Vision.Image.Storage.DevIL.ImageType

Description

Defines image types which can be used with the DevIL API.

Documentation

class ImageType t where

Methods

toIlType :: t -> Word32

Returns the DevIL constant associated with the image type.

data Autodetect

Tries to determine automatically the image type with the header and/or the file extension.

Raises an UnknownFileType error if the detection failed.

Constructors

Autodetect

data BLP

Constructors

BLP

data BMP

Constructors

BMP

data CHEAD

Constructors

CHEAD

data CUT

Constructors

CUT

data DCX

Constructors

DCX

data DDS

DirectDraw Surface (.dds).

Constructors

DDS

data DICOM

Constructors

DICOM

data Doom

Doom texture.

Constructors

Doom

data DoomFlat

Doom flat texture (floor).

Constructors

DoomFlat

data DPX

Constructors

DPX

data EXR

Constructors

EXR

data FITS

Constructors

FITS

data FTX

Constructors

FTX

data GIF

Constructors

GIF

data HDR

Constructors

HDR

data ICO

Constructors

ICO

data ICNS

Constructors

ICNS

data IFF

Constructors

IFF

data IWI

Constructors

IWI

data JASCPAL

Paint Shop Pro (Jasc) palette.

Constructors

JASCPAL

data JP2

JPEG 2000.

Constructors

JP2

data JPG

Constructors

JPG

data LIF

Homeworld (.lif).

Constructors

LIF

data MDL

Constructors

MDL

data MNG

Constructors

MNG

data MP3

Constructors

MP3

data PCD

Constructors

PCD

data PCX

Constructors

PCX

data PIC

Constructors

PIC

data PIX

Constructors

PIX

data PNG

Constructors

PNG

data PNM

Portable AnyMap (.pbm, .pgm or .ppm).

Constructors

PNM

data PSD

Photoshop Document.

Constructors

PSD

data PSP

Paint Shop Pro image.

Constructors

PSP

data PXR

Constructors

PXR

data RAW

Raw data with a 13-byte header.

Constructors

RAW

data ROT

Constructors

ROT

data SGI

Constructors

SGI

data SUN

Constructors

SUN

data Texture

Medieval II: Total War Texture (.texture) file.

Constructors

Texture

data TGA

Constructors

TGA

data TIFF

Constructors

TIFF

data TPL

Constructors

TPL

data UTX

Constructors

UTX

data VTF

Constructors

VTF

data WAL

Constructors

WAL

data WBMP

Constructors

WBMP

data XPM

Constructors

XPM